REFACTOR: AuditFields에서 createby, updatedby 컬럼 관련 로직을 auditfields를 사용하는 entity에 넘김. - setter를 설정해, 생성될때와 수정될때 auditfields의 객체 필드를 set 할 수 있도록 설정함. - aggregate/embeddable/ 에 따로 embeddable 클래스들을 관리함으로써 유지보수성을 높임
