+++ src/main/java/com/conk/integration/command/domain/aggregate/ChannelOrderItem.java
@@ -1,5 +1,7 @@
 package com.conk.integration.command.domain.aggregate;
 
+import com.conk.integration.command.domain.aggregate.embeddable.AuditFields;
+import com.conk.integration.command.domain.aggregate.embeddable.ChannelOrderItemId;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -31,5 +33,6 @@ public class ChannelOrderItem {
     private int packedQuantity = 0;
 
     @Embedded
-    private AuditFields audit;
+    @Builder.Default
+    private AuditFields audit = new AuditFields();
 }
